Page Manager

+++ Inventory.DataObjects/EDM/Item.cs
@@ -23,7 +23,7 @@ namespace Inventory.DataObjects.EDM
         public int UnitID { get; set; }
         public Nullable<int> UnitAmount { get; set; }
         public Nullable<decimal> Price { get; set; }
-        public Nullable<decimal> Quantity { get; set; }
+        public Nullable<int> Quantity { get; set; }
         public string Description { get; set; }
         public Nullable<System.DateTime> ReceivedOn { get; set; }
         public int CategoryID { get; set; }
